requirements.txt
python-telegram-bot==22.0
SQLAlchemy==2.0.20
python-dotenv==1.0.0
requests==2.31.0